+++ _examples/bootstrapper/bootstrap/bootstrapper.go
@@ -108,7 +108,7 @@ func (b *Bootstrapper) Bootstrap() *Bootstrapper {
 
 	// static files
 	b.Favicon(StaticAssets + Favicon)
-	b.HandleDir(StaticAssets[1:len(StaticAssets)-1], StaticAssets)
+	b.HandleDir("/public", iris.Dir(StaticAssets))
 
 	// middleware, after static files
 	b.Use(recover.New())