+++ _examples/file-server/embedding-files-into-app/main.go
@@ -5,8 +5,8 @@ import (
 )
 
 // Follow these steps first:
-// $ go get -u github.com/go-bindata/go-bindata/...
-// $ go-bindata ./assets/...
+// $ go get -u github.com/go-bindata/go-bindata/v3/go-bindata
+// $ go-bindata -prefix "assets" -fs ./assets/...
 // $ go run .
 // "physical" files are not used, you can delete the "assets" folder and run the example.
 //
@@ -15,13 +15,21 @@ func newApp() *iris.Application {
 	app := iris.New()
 	app.Logger().SetLevel("debug")
 
-	app.HandleDir("/static", "./assets", iris.DirOptions{
-		Asset:      Asset,
-		AssetInfo:  AssetInfo,
-		AssetNames: AssetNames,
-		ShowList:   true,
-	})
+	app.HandleDir("/static", AssetFile())
 
+	/*
+		Or if you need to cache them inside the memory (requires the assets folder
+		to be located near the executable program):
+		app.HandleDir("/static", http.Dir("./assets"), iris.DirOptions{
+			IndexName: "index.html",
+			Cache: iris.DirCacheOptions{
+				Enable:          true,
+				Encodings:       []string{"gzip"},
+				CompressIgnore:  iris.MatchImagesAssets,
+				CompressMinSize: 30 * iris.B,
+			},
+		})
+	*/
 	return app
 }